Re-orient RFC 1 to generating static config
The original idea for Spresm was for the operations to be build and
generate; i.e., it was just about making and comsuming images
containing configuration-as-code.

It's a better idea though, if it can be accomplished, to evaluate the
runnable packages and commit the output. This commit adapts the
establishing RFC so it talks about the operations of evaluating and
merging configurations as the plumbing.