Implement three-way merge of resources
kyaml has a procedure for three-way merge of resources, and for
merging three sets of files.

However, the latter is exposed as a filter that assumes it's dealing
with files, and the algorithm itself is inaccessible. It also doesn't
consider missing vs changed as a conflict, which I would like to do.

This commit introduces a procedure to do the coarse merge of sets of
resources, with the per-resource merge done by kyaml.