Get three versions in 3-way merge in order
There are three versions of the config in play when it comes to
updating:

 - the config as previously specified
 - the config as now generated
 - the config as it exists in the working directory

In a three-way merge, you are typically taking changes between a
common ancestor ("orig") and an upstream branch ("yours"), and adding
those changes to a local branch ("mine").

The orig is reasonably obviously the config as previously
specified. Before this commit, "mine" was taken to be the config as
now generated (and "yours" as the config in the working directory).

However, it makes more sense to think of this as a rebase of the
config in the working directory on upstream changes; and as such,
"mine" should be the working directory, and "yours" the newly
generated config.