Add functionConfig for image imports
This adds a config in the spec for images. I've chosen for now to use
the kyaml convention of using `functionConfig` to identify the value
given as input to an image.

This also adds `yaml` tags to the spec struct -- the behaviour of
using json tags to control YAML decoding and encoding is from the
ghodss/yaml library, but unfortunately that doesn't support
NewDecoder/NewEncoder. When using kyaml/yaml to decode and encode, you
need `yaml` tags.