Make sure redundant files are removed
Files that no longer have resources assigned to them should be
removed. You have to use the `LocalPackageReadWriter` to read the
files for this to work, so use the same one of those whether or not
--overwrite is true.

This also prints messages advising the use of --overwrite when it
appears that there's no git repo, or base revision, to use for
merging.